Rename MintDrawing.jsx to MintModal.jsx
